requirements.txt
torch==1.13.1
torchvision==0.14.1
gradio==3.16.2
Pillow>=8.0.0
numpy<2